setup.sh: don't link .gitignore

+++ setup.sh
@@ -4,7 +4,10 @@ find . -maxdepth 1 -name '.*' | while read f
 do
     f=$(basename $f)
     dest=$HOME/$f
-    if [ "$f" != ".git" ] && [ ! -f "$dest" ] && [ ! -d "$dest" ]
+    if [ "$f" != ".git" ] && \
+       [ "$f" != ".gitignore" ] && \
+       [ ! -f "$dest" ] &&   \
+       [ ! -d "$dest" ]
     then
         ln -s $PWD/$f $dest
     fi
